requirements.txt
python-telegram-bot==20.7
sqlalchemy
requests
beautifulsoup4